refactor(messaging): process server related events in factory function instead of server listener
