+++ packages/messaging/src/server/messaging.server.interface.ts
@@ -1,4 +1,4 @@
-import { BoundDependency } from '@marblejs/core';
+import { BoundDependency, ListenerServer } from '@marblejs/core';
 import { MsgServerEffect } from '../effects/messaging.effects.interface';
 import { TransportStrategy, TransportLayerConnection } from '../transport/transport.interface';
 import { messagingListener } from './messaging.server.listener';
@@ -14,6 +14,4 @@ export type CreateMicroserviceConfig =
   & ConfigurationBase
   ;
 
-export interface Microservice {
-  (): Promise<TransportLayerConnection>;
-}
+export interface Microservice extends ListenerServer<TransportLayerConnection> {}
